+++ JestDotnet/JestDotnet/Core/Serializer.cs
@@ -1,3 +1,6 @@
+using System.IO;
+using System.Text;
+using System.Text.Json;
 using JestDotnet.Core.Settings;
 
 namespace JestDotnet.Core;
@@ -6,13 +9,16 @@ internal static class Serializer
 {
     internal static string Serialize(object obj)
     {
-        var jsonSerializer = SnapshotSettings.CreateJsonSerializer();
-        using var jsonWriter = SnapshotSettings.CreateJTokenWriter();
-        jsonSerializer.Serialize(jsonWriter, obj);
-        using var stringWriter = SnapshotSettings.CreateStringWriter();
-        using var jsonTextWriter = SnapshotSettings.CreateTextWriter(stringWriter);
-        jsonWriter.Token!.WriteTo(jsonTextWriter);
-
-        return stringWriter.ToString();
+        var options = SnapshotSettings.CreateSerializerOptions();
+        using var stream = new MemoryStream();
+        using var writer = new Utf8JsonWriter(stream, new JsonWriterOptions
+        {
+            Indented = options.WriteIndented,
+            NewLine = SnapshotSettings.NewLine,
+            Encoder = options.Encoder,
+        });
+        JsonSerializer.Serialize(writer, obj, options);
+        writer.Flush();
+        return Encoding.UTF8.GetString(stream.ToArray());
     }
 }